LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_pointerleave_after_pointercancel_touch-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
pointerleave after pointercancel

Test Description: This test checks if pointerleave event triggers after pointercancel. Start touch on the black rectangle and move your touch to scroll in any direction.

Note: this test is for touch devices only

Pointer Events pointerleave tests

The following pointer types were detected: (none).


Harness Error (FAIL), message = Unhandled rejection: unimplemented

NOTRUN pointerleave event received
